Refactor: models and services
